refactor: replace custom memory allocator with std::vector in padCode
